Move the ens-object folder to shared/state
