+++ src/ensembl/src/content/app/browser/track-panel/track-panel-list/TrackPanelList.tsx
@@ -5,7 +5,10 @@ import get from 'lodash/get';
 import { toggleDrawer, changeDrawerView } from '../../drawer/drawerActions';
 import { TrackSet, BrowserTrackStates } from '../trackPanelConfig';
 import { GenomeTrackCategory } from 'src/shared/state/genome/genomeTypes';
-import { EnsObjectTrack, EnsObject } from 'src/ens-object/ensObjectTypes';
+import {
+  EnsObjectTrack,
+  EnsObject
+} from 'src/shared/state/ens-object/ensObjectTypes';
 import { RootState } from 'src/store';
 import { getIsDrawerOpened } from '../../drawer/drawerSelectors';
 import { getLaunchbarExpanded } from 'src/header/headerSelectors';
