+++ src/ensembl/src/shared/state/ens-object/ensObjectHelpers.ts
@@ -1,4 +1,4 @@
-import { EnsObject } from 'src/ens-object/ensObjectTypes';
+import { EnsObject } from 'src/shared/state/ens-object/ensObjectTypes';
 
 export const getDisplayStableId = (ensObject: EnsObject) =>
   ensObject.versioned_stable_id || ensObject.stable_id || '';
